move test to non-race to support audit assertions
